pyproject.toml
[project]
name = "telegram-weather"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28",
    "apscheduler>=3.11,<4",
    "python-dotenv>=1.0",
    "pytz>=2024.1",
]

[tool.uv]
package = false
